+++ instrumentation/r2dbc-postgresql-0.9.2/src/main/java/io/r2dbc/postgresql/client/ReactorNettyClient_Instrumentation.java
@@ -3,14 +3,19 @@ package io.r2dbc.postgresql.client;
 import com.newrelic.api.agent.weaver.MatchType;
 import com.newrelic.api.agent.weaver.NewField;
 import com.newrelic.api.agent.weaver.Weave;
+import io.r2dbc.postgresql.client.ConnectionSettings;
 import reactor.netty.Connection;
 
+import java.util.Optional;
+import java.util.TimeZone;
+
 @Weave(type = MatchType.ExactClass, originalName = "io.r2dbc.postgresql.client.ReactorNettyClient")
-public class ReactorNettyClient_Instrumentation {
+public abstract class ReactorNettyClient_Instrumentation {
     @NewField
     public final Connection clientConnection;
 
     private ReactorNettyClient_Instrumentation(Connection connection, ConnectionSettings settings) {
         this.clientConnection = connection;
     }
+
 }
