+++ settings.gradle
@@ -277,7 +277,6 @@ include 'instrumentation:r2dbc-mariadb'
 include 'instrumentation:r2dbc-mysql'
 include 'instrumentation:r2dbc-postgresql-0.9.0'
 include 'instrumentation:r2dbc-postgresql-0.9.2'
-include 'instrumentation:r2dbc-postgresql-1.0.0'
 include 'instrumentation:r2dbc-mssql'
 include 'instrumentation:rabbit-amqp-2.7'
 include 'instrumentation:rabbit-amqp-3.5.0'
